Fix project share urls using the api url instead of frontend url
Make the (i) icons display the url / token that you will copy by pressing the button
